Added history tracking option. This lets you track History for any given object and create a new History object anytime that object is changed. Anytime the save() function is called, it automatically creates the History object. This contains the "new" object in a simplified JSON form. If you use a lookup, batchLookup, scan, or query function, this object will also contain the "original" object.
Additional parameters may be passed along with a save operation,
Typically this will include things like the User that made the change,
as well as a Comment. You can also include a "transaction_id", which,
when combined with a GSI, can be used to determine all changes that
happened in a given "transaction".

To enable this functionality, you must first create a DynamoDB table
with HashKey of "obj" (S) and RangeKey of "ts" (N). You should also add
a GSI of "transaction_id" if you wish to use that feature.

Then you can enable History tracking on each individual object you wish
to have tracked, by simply adding "track_history: true" to the
definitions:

	db.create({
		track_history: true

Each object will also need a $type and $id so you can properly reference
the object that was changed.