Added "FileProperty".
The new FileProperty allows you to upload files directly to S3. It
assumes you will be doing this via the S3 POST method in your forms. To
that end, there is a new "getUploadMetadata" function, which takes an
object and a callback, and returns the information required for you to
create a POST form